Use int32 consistently for check result's ValidUseCount field. (#1154)
Former-commit-id: dfbbb55d67745d1a4f525346067fda9570524426

+++ mixer/pkg/adapter/check.go
@@ -32,7 +32,7 @@ type CheckResult struct {
 	// ValidDuration represents amount of time for which this result can be considered valid.
 	ValidDuration time.Duration
 	// ValidUseCount represents the number of uses for which this result can be considered valid.
-	ValidUseCount int64
+	ValidUseCount int32
 }
 
 // GetStatus gets status embedded in the result.
